requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch==2.0.1
torchvision==0.15.2
opencv-python==4.7.0.72
pillow==9.5.0
python-multipart==0.0.6
numpy==1.24.3
detectron2@git+https://github.com/facebookresearch/detectron2.git@v0.6
pyyaml==6.0
fvcore==0.1.5.post20221221